+++ lib/src/constants.dart
@@ -23,6 +23,9 @@ class Constants {
   static const Color offWhite = Color(0xfff0f0f0);
   static const Color headerBackground = Color(0xFFDCF0FF);
 
+  /// Time spacing for displaying initial and end time for day and week view
+  static const double initialTimeSpacing = 10;
+
   static Color get randomColor {
     return Color.fromRGBO(_random.nextInt(_maxColor),
         _random.nextInt(_maxColor), _random.nextInt(_maxColor), 1);